MergeModel with read and write in terminal
To run this file, the command should look like:
node /Users/wangyilin/Desktop/mergeModel.js 0 /Users/wangyilin/Desktop/inputModel1.json /Users/wangyilin/Desktop/inputModel2.json

/*
The first word should be 'node'.
Then, 'the path to this file' should follow.
Then, there should be the 'delta value'.
After that, the path to the json file of model 1.
Then, the path to the json file of model 2.
The output of this merge is stored in a txt file 'OutputForMerge.txt'
*/