Reset graph UI when clearing
Fixes #321. I did this by moving the removeSlider() function into the revertNodeValuesToInitial() function, and calling it inside the functiosn triggered by clear results and/or configs. I looked at use cases for revertNodeValuesToInitial() and noticed that it was always called in tandem with removeSlider() so thought it would be more efficient to move it directly inside that function.