Fixed the end placeholder bug
I deleted lines 239-246 because they saved end as "null" which created the problem. Line 239 showed end as null in the canvas when the user only selects a begin and lines 245-46 show SAVED when the user only selects a begin. I thought I cleared this error in my last debug but apparently I didn't.