Fixed issue #389  (#400)
* issue #375 And/Or end selections fixed

For issue #375, when the begin is and/or, the program used to make the end "no relationship." Now it is fixed, and users can select from template A.

* issue #375 And/Or end selections fixed

For issue #375, when the begin is and/or, the program used to make the end "no relationship." Now it is fixed, and users can select from template A.

* Update LinkInspector.js

* Issue #374 fixed in LinkInspector

Now, when begin is and/or the end will have template A, before that it will automatically assign end as "NO"

* Update LinkInspector.js

* LinkInspector.js issue #374 fixed

I fixed the bug that does not allow users to choose an end relationship when they select and/or as the begin.

* Responded to comments

* Issue #389 Fixed

Before, when the begin and end are AND/OR, switching back to Evolving Relationship would make the end "null." I fixed this issue.

* Added this.cell.prop("link-type", begin + "|" + end); back

* Fixed some overlapping changes

Because I created this branch off of another issue, the changes I made for that issue are in my code, so I just deleted them.

* Fixed the end placeholder bug

I deleted lines 239-246 because they saved end as "null" which created the problem. Line 239 showed end as null in the canvas when the user only selects a begin and lines 245-46 show SAVED when the user only selects a begin. I thought I cleared this error in my last debug but apparently I didn't.

* Revert "Fixed the end placeholder bug"

This reverts commit f1e907e9ee70b2822731110e43508112b7db9aa8.

* Fixed the end placeholder bug

I deleted lines 239-246 because they saved end as "null" which created the problem. Line 239 showed end as null in the canvas when the user only selects a begin and lines 245-46 show SAVED when the user only selects a begin. I thought I cleared this error in my last debug but apparently I didn't.