Fixed the position of the toolbar and title logo and a tempoary fix to the logo overlapping issue.
To fix the issue with the logo overlapping the configs, I made configInspector a little bit shorter. I am not sure if this is a good solution or not. I also decided to keep the logo "BloomingLeaf" seperate from the toolbar. I fixed its position as well as the toolbar.