Save and Load initial model (#489)
Modified Save/Load function to take in backbone models. Currently working for models and config, but waiting for results to show up to test if the save load works for results as well.

* Modified the LinkInspector and LinkBBM to better fit save/load

* Save Progress

* attempt to fix loading issue

See if this helps, if not we can pop off the commit. Note I also made a change in onFunctions.js, don't know how important that is. Sorry this is so messy!

* Link is now upload-able

* Temporary fix of error from JSON.Stringify() and successful upload of elements

Need to understand more of userEvaluationList in order to continue

* Fixed Function-type bug

Previously the function-type wouldn't update to what it was originally at before save

* Implemented load for ConfigCollection

* Added Comments

* Responded to comment

* Structure modification for Save Model

* Responded to comments

* Responded to Comment

* Third Save option now working

* Third Save option now working

* Revert "Third Save option now working"

This reverts commit a9702d510349050d631c588e5a0910cd36d24826.

* Third Save works now

* Updated quotes and added TODO for the future.

* Added TODO.

* Added To Do

Co-authored-by: leshghi <43756305+leshghi@users.noreply.github.com>
Co-authored-by: A. M. Grubb <amgrubb@smith.edu>