Updating the visual aspects of the toolbar and sidebar (#487)
* Fixed issue #744

Previously, the config sidebar would not scroll down. Instead, it would keep going downwards everytime a new request is made. It is not exactly issue #744, as the issue changed once the "Started Configurations Revamp in new branch" PR was mereged.

* Fixed the position of the toolbar and title logo and a tempoary fix to the logo overlapping issue.

To fix the issue with the logo overlapping the configs, I made configInspector a little bit shorter. I am not sure if this is a good solution or not. I also decided to keep the logo "BloomingLeaf" seperate from the toolbar. I fixed its position as well as the toolbar.

* Made .inspector and .attribuation disappear in analysismode

* Removed toolbar shadow

* Changed the position of the attribution to be relative.

* Fixed the placement of BloomingLeaf label

I split the text from the box style.css so I could easily change the position of the text. I am wondering if it is aceptable to manually calculate the position or to use percentage (which is a cleaner practice).

* Created space between configs and fixed some minor issues

* Moved btn-svg to save dropdown

* Simple changes to the layout

Removed the label for view mode
Changed the size of paper to fit modelling view
Fixed the length difference between the two sidebars

* Change the order of buttons in toolbar

* Fixed the issue with EVO placement

* Changed the toolbar

* Update index.html

* small change to element inspector

* add back in right inspector

* Implemented comments

Co-authored-by: A. M. Grubb <amgrubb@smith.edu>