Fixed Issues With EVO (#521)
* Fixed EVO.refresh and updated the structure for methods in EVO

We created selectResult and passed it as a para. We also updated the code to backbone.

* Fixed issue with configInspector

Before, configinspector would show up in modelling view. This is because we forgot to pass analysisresult in IntentionColoring.refresh(); line 462 evoObjects.js

* Fixed issue with EVO in modelling view

Before, EVO did not work in modelling view and when switching back and EVO was "on" it would show the colors of % and not "on." We created a varaible in the scope and passed it as parameter named viewMode.

* Fixed issues and deleted viewMode

We fixed the glitching issues and deleted the viewMode variable.

* Now when config is clicked the modelling silder is displayed

Before, when config is clicked the EVO for analysis is displayed.

* Implemented @YesugenB's comments

* Fixed elements[i]

* Updated name for IntentionColorVis

Co-authored-by: Domily11 <47505178+Domily11@users.noreply.github.com>
Co-authored-by: A. M. Grubb <amgrubb@smith.edu>