EVO-related issues with satvalue and slider (#532)
* fixed issue #221

* Removed unnecessary method

I created the update method in the last PR and forgot to delete it since we don't call it in the code.

* Fixed merge problems

* Fixed issue with select and slider

* Fixed issue with satvalue and revertNodeValuesToInitial

* Fixed issue with EVO coloring mode when creating new config

Before, when creating a new config, the EVO goes off but the EVO mode view is still on.

* Fixed issue with EVO coloring

Before, when switching from one config to the other, the color mode would be for the sliderOption before the selected sliderOption. So if it is at "satae" it shows the oloring system for "%"" etc

* Implemented Yesugen's comments and deleted chnages to `revertNodeValuesToInitial()`

* Fixed EVO Coloring Bug for State (#526)

* Fixed #525

* Responded to comments

* Reimplemented revertNodeValuesToInitial()

* Integrated IntentionColoring into EVO

* Add some docstring

* Implemented comments

Co-authored-by: Manarnz <83615524+Manarnz@users.noreply.github.com>
Co-authored-by: A. M. Grubb <amgrubb@smith.edu>

* Fixed second bug in the comment

Created a refreshSlider method to fix the issue for when the coloring is not updated while switching between config and request. Additionaly, I removed unnecessary refresh and the method EVO.update

* Cleaned code and added comments

* Made sliderShow when result is clicked and sliderOption is off

Co-authored-by: Domily11 <47505178+Domily11@users.noreply.github.com>
Co-authored-by: A. M. Grubb <amgrubb@smith.edu>