Sets it back to none
I originally removed this line because whenever you set the func type to Constant the sat val is automatically set to be none but will continue to reset when called and so overrides the previous wanted sat val. But we still need it to be set the first time at least.